1_foundations/requirements.txt
requests
python-dotenv
gradio
pypdf
PyPDF2
anthropic